Moving Thumbnail code from Workers to Tasks, Jobs
Fixing the workers table not showing when a Worker is rendering
Added Activity Collumn showing las activity on Worker
