+++ brender/manager/application/modules/workers/__init__.py
@@ -1,6 +1,4 @@
-import os
 import logging
-import requests
 
 from flask import jsonify
 from flask import request
@@ -8,17 +6,13 @@ from flask import request
 from flask.ext.restful import Resource
 from flask.ext.restful import reqparse
 from flask import Flask, request, redirect, url_for
-from werkzeug import secure_filename
 
 from application import app
 from application import db
 
 from application.modules.workers.model import Worker
 from application.modules.settings.model import Setting
-from application.modules.tasks.model import Task
 from application.helpers import http_request
-
-from threading import Thread
 
 parser = reqparse.RequestParser()
 parser.add_argument('port', type=int)
@@ -27,9 +21,7 @@ parser.add_argument('system', type=str)
 
 status_parser = reqparse.RequestParser()
 status_parser.add_argument("status", type=str)
-
-parser_thumbnail = reqparse.RequestParser()
-parser_thumbnail.add_argument("task_id", type=int)
+status_parser.add_argument("activity", type=str)
 
 class WorkerListApi(Resource):
     def post(self):
@@ -44,6 +36,7 @@ class WorkerListApi(Resource):
                           ip_address=ip_address,
                           port=port,
                           status='enabled',
+                          activity='',
                           connection='online',
                           system=args['system'])
         else:
@@ -64,6 +57,7 @@ class WorkerListApi(Resource):
             workers[worker.hostname] = {"id": worker.id,
                                         "hostname": worker.hostname,
                                         "status": worker.status,
+                                        "activity": worker.activity,
                                         "connection": worker.connection,
                                         "system": worker.system,
                                         "port" : worker.port,
@@ -78,6 +72,7 @@ class WorkerApi(Resource):
         args = status_parser.parse_args()
         worker = Worker.query.get_or_404(worker_id)
         worker.status = args['status']
+        worker.activity = args['activity']
         db.session.add(worker)
         db.session.commit()
         return jsonify(dict(status=worker.status))
@@ -85,37 +80,3 @@ class WorkerApi(Resource):
     def get(self, worker_id):
         worker = Worker.query.get_or_404(worker_id)
         return http_request(worker.host, '/run_info', 'get')
-
-class ThumbnailListApi(Resource):
-    """Thumbnail list interface for the Manager
-    """
-
-    def send_thumbnail(self, server_url, file_path, params):
-            thumbnail_file = open(file_path, 'r')
-            requests.post(server_url, files={'file': thumbnail_file}, data=params)
-            thumbnail_file.close()
-
-    def allowed_file(self, filename):
-        """Filter extensions acording to THUMBNAIL_EXTENSIONS configuration.
-        """
-        return '.' in filename and \
-               filename.rsplit('.', 1)[1] in app.config['THUMBNAIL_EXTENSIONS']
-
-    def post(self):
-        """Accepts a thumbnail file and a task_id (worker task_id),
-        and send it to the Server with the task_id (server task_id).
-        """
-
-        args = parser_thumbnail.parse_args()
-        task = Task.query.get(args['task_id'])
-        file = request.files['file']
-        full_path = os.path.join(app.config['TMP_FOLDER'], file.filename)
-        if file and self.allowed_file(file.filename):
-            filename = secure_filename(file.filename)
-            file.save(full_path)
-
-        params = dict(task_id=task.server_id)
-        server_url = "http://%s/thumbnails" % (app.config['BRENDER_SERVER'])
-
-        request_thread = Thread(target=self.send_thumbnail, args=(server_url, full_path, params))
-        request_thread.start()
