+++ brender/server/application/modules/workers/__init__.py
@@ -1,12 +1,10 @@
 import os
-import shutil
 import json
 import requests
 from flask import Flask
 from flask import jsonify
 from flask import request
 from flask import redirect
-from flask import Response
 from flask.ext.restful import Resource
 from flask.ext.restful import reqparse
 from werkzeug import secure_filename
@@ -21,9 +19,6 @@ from application.modules.tasks.model import Task
 parser = reqparse.RequestParser()
 parser.add_argument("id", type=str)
 parser.add_argument("status", type=str)
-
-parser_thumbnail = reqparse.RequestParser()
-parser_thumbnail.add_argument("task_id", type=int)
 
 class WorkerListApi(Resource):
     def get(self):
@@ -53,52 +48,3 @@ class WorkerApi(Resource):
         worker = Worker.query.get_or_404(worker_id)
         r = requests.get('http://' + worker.ip_address + '/run_info')
         return r.json()
-
-
-class ThumbnailListApi(Resource):
-    """Thumbnail list interface for the Server
-    """
-    def allowed_file(self, filename):
-        """Filter extensions acording to THUMBNAIL_EXTENSIONS configuration.
-        """
-        return '.' in filename and \
-               filename.rsplit('.', 1)[1] in app.config['THUMBNAIL_EXTENSIONS']
-
-    def post(self):
-        """Accepts a thumbnail file and a task_id and stores it.
-        """
-        args = parser_thumbnail.parse_args()
-        task = Task.query.get(args['task_id'])
-        thumbnail_filename = "thumbnail_%s.png" % task.job_id
-
-        file = request.files['file']
-        if file and self.allowed_file(file.filename):
-            filepath=os.path.join( app.config['TMP_FOLDER'] , thumbnail_filename)
-            filepath_last=os.path.join( app.config['TMP_FOLDER'] , 'thumbnail_0.png')
-            file.save(filepath)
-            shutil.copy2(filepath, filepath_last)
-
-
-class ThumbnailApi(Resource):
-    """Thumbnail interface for the Server
-    """
-    def get(self, job_id):
-        """Returns the last thumbnail for the Job, or a blank
-        image if none. If job_id is 0 return the global last
-        thumbnail.
-        """
-        def generate():
-            filename='thumbnail_%s.png' % job_id
-            file_path = os.path.join(app.config['TMP_FOLDER'],filename)
-            if os.path.isfile(file_path):
-                thumb_file = open(str(file_path), 'r')
-                return thumb_file.read()
-            else:
-                with app.open_resource('static/missing_thumbnail.png') as thumb_file:
-                    return thumb_file.read()
-            return False
-        bin = generate()
-        if bin:
-            return Response(bin, mimetype='image/png')
-        else:
-            return '',404
